Material eliminado y material1 arreglado

+++ eclipse-workspace/spring-boot-backend-apirest/src/main/java/com/tfg/springboot/backend/apirest/models/dao/IMaterialDao.java
@@ -2,10 +2,8 @@ package com.tfg.springboot.backend.apirest.models.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.tfg.springboot.backend.apirest.models.entity.Cliente;
 import com.tfg.springboot.backend.apirest.models.entity.Material;
 
-
-public interface IMaterialDao extends JpaRepository <Material, Long>{
+public interface IMaterialDao extends JpaRepository<Material, Integer> {
 
 }
